feat: Restructure examples into separate folders and packages for copy-paste convenience
